+++ src/base/Triple.cpp
@@ -96,9 +96,9 @@ std::istream& operator>>(std::istream& is, Triple<NUM>& p)
 }
 
 // Explicit instantiations.
-template class Triple<double>;
-template std::ostream& operator<<(std::ostream& os, const Triple<double> &p);
-template std::istream& operator>>(std::istream& is, Triple<double>& p);
+template class Triple<float>;
+template std::ostream& operator<<(std::ostream& os, const Triple<float> &p);
+template std::istream& operator>>(std::istream& is, Triple<float>& p);
 
 template class Triple<int>;
 template std::ostream& operator<<(std::ostream& os, const Triple<int> &p);
